Code cleanup and refactoring
