--- GPTPullRequestReview/review.ts
@@ -1,243 +1,0 @@
-import tl = require('azure-pipelines-task-lib/task');
-import fetch = require('node-fetch');
-import simpleGit = require('simple-git');
-import binaryExtensions = require('binary-extensions');
-const { Configuration, OpenAIApi } = require("openai");
-const https = require("https");
-
-const gitOptions: Partial<simpleGit.SimpleGitOptions> = {
-  baseDir: `${tl.getVariable('System.DefaultWorkingDirectory')}`,
-  binary: 'git'
-};
-
-let openai: any;
-let azureOpenai: any;
-let git: simpleGit.SimpleGit;
-let targetBranch: string;
-let httpsAgent: any;
-var apiKey: any;
-var aoiEndpoint: any;
-
-async function run() {
-  try {
-    if (tl.getVariable('Build.Reason') !== 'PullRequest') {
-      tl.setResult(tl.TaskResult.Skipped, "This task should be run only when the build is triggered from a Pull Request.");
-      return;
-    }
-
-    const supportSelfSignedCertificate = tl.getBoolInput('support_self_signed_certificate');
-    apiKey = tl.getInput('api_key', true);
-    aoiEndpoint = tl.getInput('aoi_endpoint');
-    
-    if (apiKey == undefined) {
-      tl.setResult(tl.TaskResult.Failed, 'No Api Key provided!');
-      return;
-    }
-
-    if (aoiEndpoint == undefined) {
-      const openAiConfiguration = new Configuration({
-        apiKey: apiKey,
-      });
-      
-      openai = new OpenAIApi(openAiConfiguration);
-    }
-
-    httpsAgent = new https.Agent({
-      rejectUnauthorized: !supportSelfSignedCertificate
-    });
-
-    git = simpleGit.simpleGit(gitOptions);
-    targetBranch = getTargetBranchName();
-
-    const filesNames = await GetChangedFiles(targetBranch);
-
-    await DeleteExistingComments();
-
-    for (const fileName of filesNames) {
-      await reviewFile(fileName)
-    }
-
-    tl.setResult(tl.TaskResult.Succeeded, "Pull Request reviewed.");
-  }
-  catch (err: any) {
-    tl.setResult(tl.TaskResult.Failed, err.message);
-  }
-}
-
-async function GetChangedFiles(targetBranch: string) {
-  await git.addConfig('core.pager', 'cat');
-  await git.addConfig('core.quotepath', 'false');
-  await git.fetch();
-
-  const diffs = await git.diff([targetBranch, '--name-only']);
-  const files = diffs.split('\n').filter(line => line.trim().length > 0);
-  const nonBinaryFiles = files.filter(file => !binaryExtensions.includes(getFileExtension(file)));
-
-  console.log(`Changed Files (excluding binary files) : \n ${nonBinaryFiles.join('\n')}`);
-
-  return nonBinaryFiles;
-}
-
-async function reviewFile(fileName: string) {
-  console.log(`Start reviewing ${fileName} ...`);
-
-  const patch = await git.diff([targetBranch, '--', fileName]);
-
-  const prompt = `
-          Act as a code reviewer of a Pull Request, providing feedback on the code changes below.
-          You are provided with the Pull Request changes in a patch format.
-          Each patch entry has the commit message in the Subject line followed by the code changes (diffs) in a unidiff format.
-          
-          As a code reviewer, your task is:
-          - Review only added, edited or deleted lines.
-          - Non changed code should not be reviewed
-          - If there's no bugs, write 'No feedback'.
-          - Use bullet points if you have multiple comments.
-          
-          Patch of the Pull Request to review:
-          ${patch}
-          `;
-
-  try {
-    let choices: any;
-
-    if (aoiEndpoint == undefined) {
-      const response = await openai.createCompletion({
-        model: "text-davinci-003",
-        prompt: prompt,
-        max_tokens: 500
-      });
-
-      choices = response.data.choices
-    }
-    else {
-      const request = await fetch.default(aoiEndpoint, {
-        method: 'POST',
-        headers: { 'api-key': `${apiKey}`, 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          max_tokens: 500,
-          messages: [{
-            role: "user",
-            content: prompt
-          }]
-        }),
-        agent: httpsAgent
-      });
-
-      const response = await request.json();
-      
-      choices = response.choices;
-    }
-
-    if (choices && choices.length > 0) {
-      const review = aoiEndpoint ? choices[0].message?.content : choices[0].text as string
-
-      if (review.trim() !== "No feedback.") {
-        await AddCommentToPR(fileName, review);
-      }
-    }
-
-    console.log(`Review of ${fileName} completed.`);
-  }
-  catch (error: any) {
-    if (error.response) {
-      console.log(error.response.status);
-      console.log(error.response.data);
-    } else {
-      console.log(error.message);
-    }
-  }
-}
-
-async function AddCommentToPR(fileName: string, comment: string) {
-  const body = {
-    comments: [
-      {
-        parentCommentId: 0,
-        content: comment,
-        commentType: 1
-      }
-    ],
-    status: 1,
-    threadContext: {
-      filePath: fileName,
-    }
-  }
-
-  const prUrl = `${tl.getVariable('SYSTEM.TEAMFOUNDATIONCOLLECTIONURI')}${tl.getVariable('SYSTEM.TEAMPROJECTID')}/_apis/git/repositories/${tl.getVariable('Build.Repository.Name')}/pullRequests/${tl.getVariable('System.PullRequest.PullRequestId')}/threads?api-version=5.1`
-
-  await fetch.default(prUrl, {
-    method: 'POST',
-    headers: { 'Authorization': `Bearer ${tl.getVariable('SYSTEM.ACCESSTOKEN')}`, 'Content-Type': 'application/json' },
-    body: JSON.stringify(body),
-    agent: httpsAgent
-  });
-
-  console.log(`New comment added.`);
-}
-
-async function DeleteExistingComments() {
-  console.log("Start deleting existing comments added by the previous Job ...");
-
-  const threadsUrl = `${tl.getVariable('SYSTEM.TEAMFOUNDATIONCOLLECTIONURI')}${tl.getVariable('SYSTEM.TEAMPROJECTID')}/_apis/git/repositories/${tl.getVariable('Build.Repository.Name')}/pullRequests/${tl.getVariable('System.PullRequest.PullRequestId')}/threads?api-version=5.1`;
-  const threadsResponse = await fetch.default(threadsUrl, {
-    headers: { Authorization: `Bearer ${tl.getVariable('SYSTEM.ACCESSTOKEN')}` },
-    agent: httpsAgent
-  });
-
-  const threads = await threadsResponse.json() as { value: [] };
-  const threadsWithContext = threads.value.filter((thread: any) => thread.threadContext !== null);
-
-  const collectionUri = tl.getVariable('SYSTEM.TEAMFOUNDATIONCOLLECTIONURI') as string;
-  const collectionName = getCollectionName(collectionUri);
-  const buildServiceName = `${tl.getVariable('SYSTEM.TEAMPROJECT')} Build Service (${collectionName})`;
-
-  for (const thread of threadsWithContext as any[]) {
-    const commentsUrl = `${tl.getVariable('SYSTEM.TEAMFOUNDATIONCOLLECTIONURI')}${tl.getVariable('SYSTEM.TEAMPROJECTID')}/_apis/git/repositories/${tl.getVariable('Build.Repository.Name')}/pullRequests/${tl.getVariable('System.PullRequest.PullRequestId')}/threads/${thread.id}/comments?api-version=5.1`;
-    const commentsResponse = await fetch.default(commentsUrl, {
-      headers: { Authorization: `Bearer ${tl.getVariable('SYSTEM.ACCESSTOKEN')}` },
-      agent: httpsAgent
-    });
-
-    const comments = await commentsResponse.json() as { value: [] };
-
-    for (const comment of comments.value.filter((comment: any) => comment.author.displayName === buildServiceName) as any[]) {
-      const removeCommentUrl = `${tl.getVariable('SYSTEM.TEAMFOUNDATIONCOLLECTIONURI')}${tl.getVariable('SYSTEM.TEAMPROJECTID')}/_apis/git/repositories/${tl.getVariable('Build.Repository.Name')}/pullRequests/${tl.getVariable('System.PullRequest.PullRequestId')}/threads/${thread.id}/comments/${comment.id}?api-version=5.1`;
-
-      await fetch.default(removeCommentUrl, {
-        method: 'DELETE',
-        headers: { Authorization: `Bearer ${tl.getVariable('SYSTEM.ACCESSTOKEN')}` },
-        agent: httpsAgent
-      });
-    }
-  }
-
-  console.log("Existing comments deleted.");
-}
-
-function getCollectionName(collectionUri: string) {
-  const collectionUriWithoutProtocol = collectionUri!.replace('https://', '').replace('http://', '');
-
-  if (collectionUriWithoutProtocol.includes('.visualstudio.')) {
-    return collectionUriWithoutProtocol.split('.visualstudio.')[0];
-  }
-  else {
-    return collectionUriWithoutProtocol.split('/')[1];
-  }
-}
-
-function getFileExtension(fileName: string) {
-  return fileName.slice((fileName.lastIndexOf(".") - 1 >>> 0) + 2);
-}
-
-function getTargetBranchName() {
-  let targetBranchName = tl.getVariable('System.PullRequest.TargetBranchName');
-
-  if (!targetBranchName) {
-    targetBranchName = tl.getVariable('System.PullRequest.TargetBranch')?.replace('refs/heads/', '');
-  }
-
-  return `origin/${targetBranchName}`;
-}
-
-run();
