+++ GPTPullRequestReview/src/index.ts
@@ -1,0 +1,60 @@
+import * as tl from "azure-pipelines-task-lib/task";
+import { Configuration, OpenAIApi } from 'openai';
+import { deleteExistingComments } from './pr';
+import { reviewFile } from './review';
+import { getTargetBranchName } from './utils';
+import { getChangedFiles } from './git';
+import https from 'https';
+
+let openai: OpenAIApi;
+let targetBranch: string;
+let httpsAgent: any;
+let apiKey: string | undefined;
+let aoiEndpoint: string | undefined;
+
+async function run() {
+  try {
+    if (tl.getVariable('Build.Reason') !== 'PullRequest') {
+      tl.setResult(tl.TaskResult.Skipped, "This task should be run only when the build is triggered from a Pull Request.");
+      return;
+    }
+
+    const supportSelfSignedCertificate = tl.getBoolInput('support_self_signed_certificate');
+    apiKey = tl.getInput('api_key', true);
+    aoiEndpoint = tl.getInput('aoi_endpoint');
+
+    if (apiKey == undefined) {
+      tl.setResult(tl.TaskResult.Failed, 'No Api Key provided!');
+      return;
+    }
+
+    if (aoiEndpoint == undefined) {
+      const openAiConfiguration = new Configuration({
+        apiKey: apiKey,
+      });
+
+      openai = new OpenAIApi(openAiConfiguration);
+    }
+
+    httpsAgent = new https.Agent({
+      rejectUnauthorized: !supportSelfSignedCertificate
+    });
+
+    targetBranch = getTargetBranchName();
+
+    const filesNames = await getChangedFiles(targetBranch);
+
+    await deleteExistingComments(httpsAgent);
+
+    for (const fileName of filesNames) {
+      await reviewFile(targetBranch, fileName, httpsAgent, apiKey, openai, aoiEndpoint)
+    }
+
+    tl.setResult(tl.TaskResult.Succeeded, "Pull Request reviewed.");
+  }
+  catch (err: any) {
+    tl.setResult(tl.TaskResult.Failed, err.message);
+  }
+}
+
+run();
