+++ GPTPullRequestReview/src/utils.ts
@@ -1,0 +1,15 @@
+import * as tl from "azure-pipelines-task-lib/task";
+
+export function getFileExtension(fileName: string) {
+  return fileName.slice((fileName.lastIndexOf(".") - 1 >>> 0) + 2);
+}
+
+export function getTargetBranchName() {
+  let targetBranchName = tl.getVariable('System.PullRequest.TargetBranchName');
+
+  if (!targetBranchName) {
+    targetBranchName = tl.getVariable('System.PullRequest.TargetBranch')?.replace('refs/heads/', '');
+  }
+
+  return `origin/${targetBranchName}`;
+}
