+++ vss-extension.json
@@ -2,7 +2,7 @@
     "manifestVersion": 1,
     "id": "GPTPullRequestReview",
     "name": "GPT Pull Request Review",
-    "version": "0.9.18",
+    "version": "0.10.2",
     "publisher": "mustaphalarhrouch",
     "galleryFlags": [
         "Public",
